Move generateIdSalt to util folder
